LayoutTests/fast/events/touch/document-create-touch-list-expected.txt

This tests support for the document.createTouchList API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "createTouchList" in document is true
PASS touchList is non-null.
PASS touchList.length is 0
PASS touchList.item(0) is null
PASS touchList.item(1) is null
PASS successfullyParsed is true

TEST COMPLETE
